+++ psw/ae/common/src/AECloseSessionRequest.cpp
@@ -79,7 +79,7 @@ AEMessage* AECloseSessionRequest::serialize()
             ae_msg = new AEMessage;
             ae_msg->size = (unsigned int)msg.ByteSize();
             ae_msg->data = new char[ae_msg->size];
-            msg.SerializeToArray(ae_msg->data, ae_msg->size);
+            msg.SerializeToArray(ae_msg->data, (int)ae_msg->size);
         }
     }
     return ae_msg;
